refactor: 음식점 목록 컴포넌트화
