+++ index.html
@@ -38,7 +38,6 @@
 
       <!-- 음식점 목록 -->
       <section class="restaurant-list-container">
-        <ul class="restaurant-list"></ul>
       </section>
 
       <!-- 음식점 추가 모달 -->
