+++ src/domain/filter.ts
@@ -1,6 +1,6 @@
 import { LOCAL_STORAGE_KEY, SELECTED_OPTION } from "../constant";
 import { RestaurantType } from "../type";
-import { renderRestaurantList } from "../ui/restaurantListRenderer";
+import { renderRestaurantList } from "../component/restaurantList";
 import { getAllRestaurantsInLocalStorage } from "./localStorageController";
 const { All_CATEGORIES, NAME, DISTANCE } = SELECTED_OPTION;
 const { CATEGORY, SORT } = LOCAL_STORAGE_KEY;
@@ -42,7 +42,6 @@ export const sortRestaurantList = (
 export const updateRestaurantList = () => {
   const filteredRestaurantList = filterCategory(
     localStorage.getItem(CATEGORY) as string
-    
   );
 
   sortRestaurantList(
